feat(layout): 项目结构调整

+++ internal/apps/demoapp/dto/dtoexample/request.go
@@ -1,5 +1,5 @@
 package dtoexample
 
 type GetUserReq struct {
-	ID uint `json:"id" form:"id"` // 用户ID
+	ID uint64 `json:"id" form:"id"` // 用户ID
 }
